Product_done

+++ app/src/main/java/com/example/slcraft/models/Calculate.kt
@@ -1,5 +1,5 @@
 package com.example.slcraft.models
 
-class Calculate (private val quantity:Double, private val price:Double){
+class Calculate (private val quantity:Int, private val price:Double){
     fun multiply() = quantity*price
 }
